+++ Common/Info/InfoMapShadow.cpp
@@ -523,7 +523,7 @@ void CInfoMapShadow::AddAnime(void)
 
 	pInfo = new CInfoAnime;
 	m_paAnimeInfo->Add (pInfo);
-	m_byAnimeCount = m_paAnimeInfo->size();
+	m_byAnimeCount = static_cast<BYTE>(m_paAnimeInfo->size());
 }
 
 
@@ -540,7 +540,7 @@ void CInfoMapShadow::DeleteAnime(int nNo)
 	pInfo = m_paAnimeInfo->at(nNo);
 	if ((nNo >= 0) && (nNo < static_cast<int>(m_paAnimeInfo->size()))) { m_paAnimeInfo->erase (m_paAnimeInfo->begin () + nNo); }
 	SAFE_DELETE (pInfo);
-	m_byAnimeCount = m_paAnimeInfo->size();
+	m_byAnimeCount = static_cast<BYTE>(m_paAnimeInfo->size());
 }
